requirements.txt
antiorm==1.2.1
asgiref==3.4.1
backports.zoneinfo==0.2.1
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
db==0.1.1
defusedxml==0.7.1
Django==3.0
django-allauth==0.47.0
django-ckeditor==6.2.0
django-js-asset==1.2.2
django-model-utils==4.2.0
django-mptt==0.13.4
django-notifications==0.1.dev0
django-notifications-hq==1.6.0
django-password-reset==2.0
django-simpleui==2022.1
django-summernote==0.8.20.0
django-taggit==2.0.0
idna==3.3
image==1.5.33
importlib-metadata==4.10.0
jsonfield==3.1.0
Markdown==3.3.6
oauthlib==3.1.1
packaging==21.3
Pillow==9.0.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.6
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
tzdata==2021.5
urllib3==1.26.8
webencodings==0.5.1
wincertstore==0.2
zipp==3.7.0
